Update echo and export functions.
This commit updates the two builtin functions echo and export.
Echo can handle '-n' as a flag input and prints the arguments
accordingly with or without a new line. The export functions can handle
multiple different things. When a variable is exported, that does not
yet exist in the environment, it will be added to the environment. If it
does exist, it will swap out the currently existing one. When exporting
something like 'export HOME' it depends if the variable exists or not.
If it already exists, it does not change anything. If it does not exist,
a new env variable will be appended with an empty string (own design
choice).