major changes regarding builtin and parse functions
A execcmd now has a new field that is holding all the attributes of the shell struct in it.
Instead of passing on the char **env variable, we now pass on the t_main *sh struct. Furthermore, export and cd
now also reflect those changes. It is working as expected.