Update Makefile and begin of mainloop
Arrange includes also to the defines.h file. Also create a new folder
called `core` which holds the main loop of our shell.
Add error checking in the parser and executor to ensure that if the
lexer failed, we will not get a segfault for that.
Add error handling for the tcgetattr function in the signal_handler to
ensure that the function does not continue when running into an error.