major changes
A lot of changes in the executor. Change the flow to better go through the AST. Forking is done
solely from the parent to keep a flat level of forking. This enables us to get the correct exit code from the last child.
Introduce also the main loop for the minishell to get the testing going. Cleanup is not completly done. Builtin 'exit' function
is also changed to adding support for checking the correct input arguments for this command.
Tasks still to do:
- Redirections + heredoc
- Segfault when unclosed quotes
- Exit code error from child
- Address issues identified by static code analyzer