more major changes
Add dup to initializing the shell for saving stdin, stdout, and stderr for resetting the file descriptors after cmd processing.
Redirection is now also working, only heredoc needs to be added from here on. exit code is not 100% how I want it to be but still
good enough. Add array of PIDs to wait for all childs after executing.
Changes needed:
- error handling in parsing (NULL pointer dereferencing)