comments, rearrange header files and executor
Added missing documentation for almost all of the functions (some in the
lexer are missing). Rearrange the inclusion of the header files to
only have the necessary files in hand and not all bundled in one. Remove
unnecessary functions and comments. Add enum for the quit check.
Declutter the executor to have them all in separated files. Make,
besides of the 42 header, all files norm conform (some comments need to
be addressed).